+++ docs/CHANGELOG.md
@@ -8,12 +8,14 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 - Support for BIDS 1.8.0
+- Test framework
 - A BIDSCOIN_DEBUG environment variable to run bidscoin in a more verbose debug mode
 - VERBOSE and SUCCESS logging levels
 - Checks using the bids-validator python module
 - Options to perform bidsmap tests with `bidscoin -t / -b`
 - Option to list and install template bidsmaps
 - The skullstrip tool (a bids-wrapper around synthstrip)
+- Support for / more robust handling of non-alphanumeric characters in sub/ses prefixes
 
 ### Changed
 - The default behaviour is now not to unzip data in the sourcefolders anymore. Instead, users can use the `--unzip` option in the bidsmapper